implement error handling

+++ Application/Activities/Comands/DeleteActivity.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Reflection.Metadata.Ecma335;
+using Application.Core;
 using AutoMapper;
 using MediatR;
 using Persistence;
@@ -7,22 +9,29 @@ namespace Application.Activities.Comands;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public required  string id { get; set; }
 
-        public class Handler(AppDbContext context) : IRequestHandler<Command>
+        public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
 
                 var activity = await context.Activites
-                .FindAsync([request.id], cancellationToken)
-                ?? throw new Exception("Cannot find activity");
+                .FindAsync([request.id], cancellationToken);
+
+                if (activity == null)
+                {
+                    return Result<Unit>.Failure("Activity not found", 404);
+                }
 
                 context.Activites.Remove(activity);
 
-                await context.SaveChangesAsync(cancellationToken);
+                var result = await context.SaveChangesAsync(cancellationToken) > 0;
+                if (!result) return Result<Unit>.Failure("Failed to delete Activity", 400);
+
+                return Result<Unit>.Success(Unit.Value);
 
             }
         }
